A lot of implementations have been modified.
