+++ lib/__init__.py
@@ -6,7 +6,7 @@
 # @File    : __init__.py.py
 # @Software: PyCharm
 # @Mail    : Jacquewong@stu.jluzh.edu.cn
-
+import src
 from lib.control import *
 from lib.custom_log import *
 from lib.config import *
@@ -14,4 +14,8 @@ from lib.resource import *
 from lib.auto import *
 from lib.evade import *
 from lib.data import *
+from lib.record import *
+
+log = src.log
+
